Consume type info and defaults from code
Add 'sphinx.ext.napoleon' and 'sphinx_autodoc_typehints' to the
documentation configuration, to consume the type information and
defaults from the code instead of duplicating it in the docstring.

Use dependency groups instead of extras in the 'pyproject.toml',
and simplify the setup by only having one development dependencies
group.