Fix test setup, update dependencies (#137)
Fix test setup on Linux by introducing a small `sleep` after launching the server. The reason why this is needed is not completely clear, I suspect a problem with binding to the port. In PyACP tests (which launch via `docker compose`), we've never seen this problem.

Improve the error handling in the helper context managers of the tests.

Update the poetry lockfile and pre-commit hooks.